Improved select-element
-reimplement select elem based on input (now availiable with Tab and IOS)
-improved code and file structure
