+++ js/select/shutDownSelect.js
@@ -1,0 +1,6 @@
+//скрывает options-container открытого селектора
+export const shutDownSelect = () => {
+  document.querySelector(".select").classList.remove("select_active");
+  document.querySelector(".input-select").disabled = "";
+  document.querySelector(".options-container").remove();
+};
